requirements.txt
aiogram==2.17.1
requests==2.26.0
python-dotenv==0.19.2
psycopg2==2.9.3
emoji==1.6.3